Style About section
Added my profile picture for the about section and
added Lorem Ipsum text as a place holder for now.
Still need to add real description for about
section and then add the technologies I know below
it. Want to add the technologies as the logos of
each one I know.