Begin button functionality for experience section
I've got rid of the button circle and added a function handle the button
click that works. I now need to make the function underline/highlight
which experience is clicked on and display my work experience from that
position. Could experiment with some animations to make it obvious which
button is selected. Maybe grow selected button and shrink other options?