Refactor Job Selector Code
I completed choosing which job is selected using
radio buttons, and changed the format because a
side by side looked weird. Am working on adding
job description below the selector next. I want
to implement the same idea as the selector for
the descriptions to make it scalable/easy to
make changes.